Added homing to SutterDebugSerialPort, cleaned and moved tests around
