+++ hardwarelibrary/tests/testCommunicationAndSerialPorts.py
@@ -243,18 +243,18 @@ def threadReadWrite(port, index):
 
 
 
-# class TestDebugEchoPort(BaseTestCases.TestEchoPort):
+class TestDebugEchoPort(BaseTestCases.TestEchoPort):
 
-#     def setUp(self):
-#         self.port = DebugEchoPort()
-#         self.assertIsNotNone(self.port)
-#         self.port.open()
-#         self.assertTrue(self.port.isOpen)
-#         self.port.flush()
+    def setUp(self):
+        self.port = DebugEchoPort()
+        self.assertIsNotNone(self.port)
+        self.port.open()
+        self.assertTrue(self.port.isOpen)
+        self.port.flush()
 
-#     def tearDown(self):
-#         self.port.close()
-#         self.assertFalse(self.port.isOpen)
+    def tearDown(self):
+        self.port.close()
+        self.assertFalse(self.port.isOpen)
 
 class TestDebugPortDefaultsToECho(BaseTestCases.TestEchoPort):
 
@@ -320,34 +320,35 @@ class TestFTDIAdaptor(unittest.TestCase):
         for c in data[:nBytes]:
             print("{0:08b} ".format(c),end='')
 
-# class TestSlowDebugEchoPort(BaseTestCases.TestEchoPort):
+class TestSlowDebugEchoPort(BaseTestCases.TestEchoPort):
 
-#     def setUp(self):
-#         self.port = DebugEchoPort()
-#         self.assertIsNotNone(self.port)
-#         self.port.open()
-#         self.assertTrue(self.port.isOpen)
-#         self.port.delay = 0.01
-#         self.port.flush()
+    def setUp(self):
+        self.port = DebugEchoPort()
+        self.assertIsNotNone(self.port)
+        self.port.open()
+        self.assertTrue(self.port.isOpen)
+        self.port.delay = 0.01
+        self.port.flush()
 
-#     def tearDown(self):
-#         self.port.close()
+    def tearDown(self):
+        self.port.close()
 
 
-# class TestRealEchoSerialPort(BaseTestCases.TestEchoPort):
+class TestRealEchoSerialPort(BaseTestCases.TestEchoPort):
 
-#     def setUp(self):
-#         try:
-#             self.port = SerialPort(idVendor=0x0403, idProduct=0x6001, serialNumber="FTDXIKC4")
-#             self.assertIsNotNone(self.port)
-#             self.port.open()
-#             self.port.flush()
-#         except Exception as err:
-#             raise unittest.SkipTest("No ECHO device connected {0}".format(err))
+    def setUp(self):
+        try:
+            self.port = SerialPort(idVendor=0x0403, idProduct=0x6001, serialNumber="FTDXIKC4")
+            self.assertIsNotNone(self.port)
+            self.port.open()
+            self.port.flush()
+        except Exception as err:
+            raise(unittest.SkipTest("No ECHO device connected {0}".format(err)))
 
-#     def tearDown(self):
-#         self.port.flush()
-#         self.port.close()
+    def tearDown(self):
+        if self.port is not None:
+            self.port.flush()
+            self.port.close()
 
 
 if __name__ == '__main__':
