+++ hardwarelibrary/tests/testSerialModule.py
